fix bugs in messages printing and in overflow in get_timestamp()

+++ srcs/philo/get_timestamp.c
@@ -6,7 +6,7 @@
 /*   By: dshatilo <dshatilo@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/29 18:09:02 by dshatilo          #+#    #+#             */
-/*   Updated: 2024/04/29 18:29:50 by dshatilo         ###   ########.fr       */
+/*   Updated: 2024/05/07 10:56:27 by dshatilo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 size_t	get_timestamp(void)
 {
 	struct timeval	curr_time;
-	int				timestamp;
+	size_t			timestamp;
 
 	gettimeofday(&curr_time, NULL);
 	timestamp = curr_time.tv_sec * 1000000 + curr_time.tv_usec;
